Extended available options for selecting data; instr method not case-sensitive
See __getitem__; can now pass an array of IDs, fields, or booleans.
Giving a single ID or list will return the astropy Table including just those events.
Giving an array of fields will return a table including only those fields for the current selection.
Providing a Boolean (of the same length as the selection) will return a Table including only the events for which the array is True