Load sources as pandas table, optionally from db
Cleaned up the definition of the Sources table, and implemented it instead
as a Pandas table. Made sure to close the FITS file after reading. Some
simple checking of the integrity of the table and the associated
information.

Also now can read in the (extended, post-DR2) source table from the
database, which otherwise should operate as for the FITS table